CTF: add all members of an enum type to psymtab
With the following change which was made last April:

    [gdb] Use partial symbol table to find language for main
    commit f36b492c7c852477af5e9deffed9eec107c6f37d

The ctf reader was modified to enter all members of an enum type,
similar to what the dwarf2 reader did, into the psymtab or gdb
won't be able to find them. In addition, the empty name checking
needed to be moved down so members of a unnamed enum were not left
out.

gdb/ChangeLog:

	* ctfread.c (ctf_psymtab_add_enums): New function.
	(ctf_psymtab_type_cb): call ctf_psymtab_add_enums.