gdb, testsuite, btrace: relax unneeded stepi expected output
In gdb.btrace/reconnect.exp, we test that we can disconnect and reconnect
again to a GDB session that is recording with the btrace recording format.
It does not really matter what we are recording.

The test assumed that stepping from _start will bring us into an area
without debug information.  This is not correct on all systems.

Relax the expected output to also support systems where we do have debug
information for that code.